requirements.txt
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.13.4
behave==1.2.6
bota==4.0.155
botasaurus==4.0.97
botasaurus-humancursor==4.0.83
botasaurus_api==4.0.10
botasaurus_driver==4.0.92
botasaurus_proxy_authentication==1.0.16
botasaurus_requests==4.0.38
brotli==1.2.0
bs4==0.0.2
cattrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.5
click==8.3.1
close_chrome==4.0.40
cloudscraper==1.2.71
colorama==0.4.6
cssselect==1.3.0
DateTime==6.0
Deprecated==1.3.1
docutils==0.22.4
et_xmlfile==2.0.0
exceptiongroup==1.3.1
execnet==2.1.1
fasteners==0.20
filelock==3.25.2
gevent==25.9.1
geventhttpclient==2.3.9
greenlet==3.3.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
javascript_fixes==1.1.29
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
loro==1.10.3
lxml==6.0.2
marimo==0.21.1
Markdown==3.10.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgspec==0.20.0
mycdp==1.3.3
narwhals==2.18.0
numpy==2.4.3
openpyxl==3.1.5
orjson==3.11.7
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
parameterized==0.9.0
parse==1.21.1
parse_type==0.6.6
parso==0.8.6
pdbp==1.8.2
platformdirs==4.9.4
pluggy==1.6.0
psutil==7.2.2
pycparser==3.0
Pygments==2.19.2
pymdown-extensions==10.21
PyMySQL==1.1.2
pynose==1.5.5
pyotp==2.9.0
pyparsing==3.3.2
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.5
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytweening==1.2.0
pytz==2026.1.post1
PyVirtualDisplay==3.0
PyYAML==6.0.3
requests==2.32.3
requests-cache==1.3.1
requests-toolbelt==1.0.0
rich==14.3.3
rootutils==1.0.7
sbvirtualdisplay==1.4.0
ScraperFC==4.3.0
selenium==4.32.0
seleniumbase==4.38.3
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
soccerdata==1.8.8
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.48
starlette==0.52.1
tabcompleter==1.4.0
tomlkit==0.14.0
tqdm==4.67.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
undetected-chromedriver==3.5.5
Unidecode==1.4.0
url-normalize==2.2.1
urllib3==1.26.20
uvicorn==0.42.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
wheel==0.46.3
wrapper-tls-requests==1.2.5
wrapt==2.1.2
wsproto==1.2.0
zope.event==6.1
zope.interface==8.2
